Use service directly
