+++ maven-core/src/main/java/org/apache/maven/internal/impl/DefaultArtifact.java
@@ -24,6 +24,7 @@ import org.apache.maven.api.Artifact;
 import org.apache.maven.api.ArtifactCoordinate;
 import org.apache.maven.api.Version;
 import org.apache.maven.api.annotations.Nonnull;
+import org.apache.maven.repository.internal.DefaultVersionParser;
 
 import static org.apache.maven.internal.impl.Utils.nonNull;
