+++ maven-core/src/main/java/org/apache/maven/internal/impl/DefaultDependency.java
@@ -30,6 +30,7 @@ import org.apache.maven.api.Version;
 import org.apache.maven.api.annotations.Nonnull;
 import org.apache.maven.api.annotations.Nullable;
 import org.apache.maven.api.services.TypeRegistry;
+import org.apache.maven.repository.internal.DefaultVersionParser;
 import org.eclipse.aether.artifact.ArtifactProperties;
 
 import static org.apache.maven.internal.impl.Utils.nonNull;
